requirements.txt
streamlit
requests
beautifulsoup4
httpx
python-dotenv
